[MS][Node Dev] Fix codechekc for  MaxPoolWithArgmaxV2

+++ tests/st/ops/ascend/test_maxpool_with_argmax_v2.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Huawei Technologies Co., Ltd
+# Copyright 2023 Huawei Technologies Co., Ltd
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -40,7 +40,7 @@ def test_maxpool_with_argmax_v2_ceil_mode_false():
     Expectation: success.
     """
     attributes = {'kernel_size': (3, 2), 'strides': (2, 1), 'pads': 0, 'dilation': 1,
-                  'ceil_mode': False, 'argmax_type': mstype.int64}
+                  'ceil_mode': False, 'argmax_type': mstype.int32}
     x = Tensor(np.arange(20 * 16 * 50 * 32).reshape((20, 16, 50, 32)), mstype.float32)
     net = MaxPoolWithArgmaxV2Net(**attributes)
     output, _ = net(x)
@@ -55,7 +55,7 @@ def test_maxpool_with_argmax_v2_ceil_mode_true():
     Expectation: success.
     """
     attributes = {'kernel_size': (3, 2), 'strides': (2, 1), 'pads': 0, 'dilation': 1,
-                  'ceil_mode': True, 'argmax_type': mstype.int64}
+                  'ceil_mode': True, 'argmax_type': mstype.int32}
     x = Tensor(np.arange(20 * 16 * 50 * 32).reshape((20, 16, 50, 32)), mstype.float32)
     net = MaxPoolWithArgmaxV2Net(**attributes)
     output, _ = net(x)
